refactor(core): Extract ProjectDeletionService to break the project service dependency cycle
The team-project deletion cascade lived in ProjectService and reached back
into WorkflowService and CredentialsService via lazy dynamic imports, while
both of those services inject ProjectService directly. That formed two
import cycles (project <-> workflow, project <-> credentials).

Move the cascade into a new ProjectDeletionService that sits above the leaf
services and calls them directly, so ProjectService no longer imports them.
The folder move now goes straight to FolderRepository. Optional-module
cleanups (data-table, external-secrets, agents) keep their lazy imports.
