+++ packages/cli/src/services/__tests__/project.service.ee.test.ts
@@ -1,5 +1,4 @@
 import type { ProjectRelation } from '@n8n/api-types';
-import type { Logger, ModuleRegistry } from '@n8n/backend-common';
 import {
 	type Project,
 	type ProjectRepository,
@@ -17,9 +16,6 @@ import type { Mocked } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
 import type { ICredentialConnectionStatusProvider } from '@/credentials/credential-connection-status-provider.interface';
-import type { AgentExecutionService } from '@/modules/agents/agent-execution.service';
-import type { AgentKnowledgeService } from '@/modules/agents/agent-knowledge.service';
-import type { AgentRepository } from '@/modules/agents/repositories/agent.repository';
 
 import type { OwnershipService } from '../ownership.service';
 import { ProjectService } from '../project.service.ee';
@@ -32,22 +28,14 @@ describe('ProjectService', () => {
 	const projectRelationRepository = mock<ProjectRelationRepository>({ manager });
 	const roleService = mock<RoleService>();
 	const sharedCredentialsRepository = mock<SharedCredentialsRepository>();
-	const moduleRegistry = mock<ModuleRegistry>({ entities: [] });
-	const agentRepository = mock<AgentRepository>();
-	const agentKnowledgeService = mock<AgentKnowledgeService>();
-	const agentExecutionService = mock<AgentExecutionService>();
 	const ownershipService = mock<OwnershipService>();
-	const logger = mock<Logger>();
 	const projectService = new ProjectService(
 		sharedWorkflowRepository,
 		projectRepository,
 		projectRelationRepository,
 		roleService,
-		sharedCredentialsRepository,
 		mock(),
-		moduleRegistry,
 		ownershipService,
-		logger,
 	);
 
 	beforeEach(() => {
@@ -524,152 +512,6 @@ describe('ProjectService', () => {
 				where: { id: projectId, type: 'team' },
 				relations: { projectRelations: { role: true } },
 			});
-		});
-	});
-
-	describe('deleteProject', () => {
-		const user = { id: 'user-1', role: { scopes: [{ slug: 'project:delete' }] } } as any;
-
-		beforeEach(() => {
-			Object.defineProperty(projectService, 'workflowService', {
-				configurable: true,
-				get: async () => ({ delete: vi.fn() }),
-			});
-			Object.defineProperty(projectService, 'credentialsService', {
-				configurable: true,
-				get: async () => ({ delete: vi.fn() }),
-			});
-		});
-
-		it('calls cleanupOrphanedEntriesForUsers with member IDs after project is deleted', async () => {
-			// ARRANGE
-			const project = mock<Project>({ id: 'project-1', type: 'team' });
-			const mockProxy = mock<ICredentialConnectionStatusProvider>();
-			Object.defineProperty(projectService, 'connectionStatusProxy', {
-				configurable: true,
-				get: async () => mockProxy,
-			});
-			manager.findOne.mockResolvedValueOnce(project);
-			projectRepository.remove.mockResolvedValueOnce(project);
-			sharedWorkflowRepository.find.mockResolvedValueOnce([]);
-			sharedCredentialsRepository.find.mockResolvedValueOnce([]);
-			moduleRegistry.isActive.mockReturnValue(false);
-			// Two members in the project
-			projectRelationRepository.findBy.mockResolvedValueOnce([
-				{ userId: 'member-1' },
-				{ userId: 'member-2' },
-			] as never);
-
-			// ACT
-			await projectService.deleteProject(user, project.id);
-
-			// ASSERT — project removed first, then cleanup for former members
-			expect(projectRepository.remove).toHaveBeenCalledWith(project);
-			expect(mockProxy.cleanupOrphanedEntriesForUsers).toHaveBeenCalledWith([
-				'member-1',
-				'member-2',
-			]);
-			expect(projectRepository.remove.mock.invocationCallOrder[0]).toBeLessThan(
-				mockProxy.cleanupOrphanedEntriesForUsers.mock.invocationCallOrder[0],
-			);
-		});
-
-		it('skips credential cleanup when the project had no members', async () => {
-			// ARRANGE
-			const project = mock<Project>({ id: 'project-1', type: 'team' });
-			const mockProxy = mock<ICredentialConnectionStatusProvider>();
-			Object.defineProperty(projectService, 'connectionStatusProxy', {
-				configurable: true,
-				get: async () => mockProxy,
-			});
-			manager.findOne.mockResolvedValueOnce(project);
-			projectRepository.remove.mockResolvedValueOnce(project);
-			sharedWorkflowRepository.find.mockResolvedValueOnce([]);
-			sharedCredentialsRepository.find.mockResolvedValueOnce([]);
-			moduleRegistry.isActive.mockReturnValue(false);
-			projectRelationRepository.findBy.mockResolvedValueOnce([]);
-
-			// ACT
-			await projectService.deleteProject(user, project.id);
-
-			// ASSERT — no members → cleanup must not be called
-			expect(projectRepository.remove).toHaveBeenCalledWith(project);
-			expect(mockProxy.cleanupOrphanedEntriesForUsers).not.toHaveBeenCalled();
-		});
-
-		it('cleans agent knowledge files before project deletion cascades agent files', async () => {
-			const project = mock<Project>({ id: 'project-1', type: 'team' });
-			Object.defineProperty(projectService, 'agentRepository', {
-				configurable: true,
-				get: async () => agentRepository,
-			});
-			Object.defineProperty(projectService, 'agentKnowledgeService', {
-				configurable: true,
-				get: async () => agentKnowledgeService,
-			});
-			Object.defineProperty(projectService, 'agentExecutionService', {
-				configurable: true,
-				get: async () => agentExecutionService,
-			});
-			manager.findOne.mockResolvedValueOnce(project);
-			projectRepository.remove.mockResolvedValueOnce(project);
-			sharedWorkflowRepository.find.mockResolvedValueOnce([]);
-			sharedCredentialsRepository.find.mockResolvedValueOnce([]);
-			moduleRegistry.isActive.mockImplementation((moduleName) => moduleName === 'agents');
-			projectRelationRepository.findBy.mockResolvedValueOnce([]);
-			agentRepository.findByProjectId.mockResolvedValueOnce([
-				{ id: 'agent-1' },
-				{ id: 'agent-2' },
-			] as never);
-
-			await projectService.deleteProject(user, project.id);
-
-			expect(agentRepository.findByProjectId).toHaveBeenCalledWith(project.id);
-			expect(agentKnowledgeService.deleteAllFilesForAgent).toHaveBeenCalledWith(
-				project.id,
-				'agent-1',
-			);
-			expect(agentKnowledgeService.deleteAllFilesForAgent).toHaveBeenCalledWith(
-				project.id,
-				'agent-2',
-			);
-			expect(agentKnowledgeService.deleteAllFilesForAgent.mock.invocationCallOrder[1]).toBeLessThan(
-				projectRepository.remove.mock.invocationCallOrder[0],
-			);
-			expect(agentKnowledgeService.destroySandbox).toHaveBeenCalledWith(project.id, 'agent-1');
-			expect(agentKnowledgeService.destroySandbox).toHaveBeenCalledWith(project.id, 'agent-2');
-			expect(agentExecutionService.deleteExecutionLogsForAgent).toHaveBeenCalledWith('agent-1');
-			expect(agentExecutionService.deleteExecutionLogsForAgent).toHaveBeenCalledWith('agent-2');
-		});
-
-		it('destroys agent sandboxes even when knowledge file cleanup fails', async () => {
-			const project = mock<Project>({ id: 'project-1', type: 'team' });
-			Object.defineProperty(projectService, 'agentRepository', {
-				configurable: true,
-				get: async () => agentRepository,
-			});
-			Object.defineProperty(projectService, 'agentKnowledgeService', {
-				configurable: true,
-				get: async () => agentKnowledgeService,
-			});
-			Object.defineProperty(projectService, 'agentExecutionService', {
-				configurable: true,
-				get: async () => agentExecutionService,
-			});
-			manager.findOne.mockResolvedValueOnce(project);
-			projectRepository.remove.mockResolvedValueOnce(project);
-			sharedWorkflowRepository.find.mockResolvedValueOnce([]);
-			sharedCredentialsRepository.find.mockResolvedValueOnce([]);
-			moduleRegistry.isActive.mockImplementation((moduleName) => moduleName === 'agents');
-			projectRelationRepository.findBy.mockResolvedValueOnce([]);
-			agentRepository.findByProjectId.mockResolvedValueOnce([{ id: 'agent-1' }] as never);
-			agentKnowledgeService.deleteAllFilesForAgent.mockRejectedValueOnce(new Error('storage down'));
-
-			await expect(projectService.deleteProject(user, project.id)).resolves.toBeUndefined();
-
-			expect(agentKnowledgeService.destroySandbox).toHaveBeenCalledWith(project.id, 'agent-1');
-			expect(agentExecutionService.deleteExecutionLogsForAgent).toHaveBeenCalledWith('agent-1');
-			expect(projectRepository.remove).toHaveBeenCalledWith(project);
 		});
 	});
